+++ shared/bpmn-designer/types/index.d.ts
@@ -27,5 +27,3 @@ declare module 'bpmnlint/rules/single-event-definition';
 declare module 'bpmnlint/rules/start-event-required';
 declare module 'bpmnlint/rules/sub-process-blank-start-event';
 declare module 'bpmnlint/rules/superfluous-gateway';
-
-declare module '@herodotus-cloud/components';