+++ shared/core/src/enums/index.ts
@@ -36,6 +36,16 @@ export enum AuthorizationGrantTypeEnum {
   TOKEN_EXCHANGE = 'urn:ietf:params:oauth:grant-type:token-exchange',
 }
 
+export enum ClientAuthenticationMethodEnum {
+  CLIENT_SECRET_BASIC = 'client_secret_basic',
+  CLIENT_SECRET_POST = 'client_secret_post',
+  CLIENT_SECRET_JWT = 'client_secret_jwt',
+  PRIVATE_KEY_JWT = 'private_key_jwt',
+  NONE = 'none',
+  TLS_CLIENT_AUTH = 'tls_client_auth',
+  SELF_SIGNED_TLS_CLIENT_AUTH = 'self_signed_tls_client_auth',
+}
+
 export enum BuildInScopeEnum {
   OPENID = 'openid',
   EMAIL = 'email',