+++ shared/core/src/lib/definitions/axios/core.ts
@@ -3,14 +3,14 @@ import type {
   AxiosResponse,
   AxiosError,
   ParamsSerializerOptions,
-  RawAxiosRequestConfig,
   InternalAxiosRequestConfig,
-  HttpResult,
-  AxiosHook,
   AxiosHttpResult,
-  AxiosHeaderStrategy,
-  AxiosRequestStrategy,
-  RequestOptions,
+  AxiosInstanceHooks,
+  AxiosRequestConfig,
+  HttpHeaderPolicy,
+  HttpRequestOptions,
+  HttpRequestPolicy,
+  HttpResult,
 } from '@/declarations';
 
 import { ContentTypeEnum, HttpMethodEnum } from '@/enums';
@@ -26,60 +26,63 @@ import { lodash } from '../../utils';
  */
 export class Axios {
   private axiosInstance: AxiosInstance;
-  private readonly axiosConfig: RawAxiosRequestConfig;
-  private readonly axiosHook: AxiosHook;
-  private readonly defaultRequestOptions: RequestOptions;
+  private readonly defaultAxiosRequestConfig: AxiosRequestConfig;
+  private readonly defaultRequestOptions: HttpRequestOptions;
+  private readonly axiosInstanceHooks: AxiosInstanceHooks;
 
-  constructor(config: RawAxiosRequestConfig, hook: AxiosHook, options: RequestOptions) {
-    this.axiosConfig = config;
-    this.axiosHook = hook;
-    this.defaultRequestOptions = options;
-    this.axiosInstance = this.createAxiosInstance(config);
+  constructor(
+    defaultAxiosRequestConfig: AxiosRequestConfig,
+    axiosInstanceHooks: AxiosInstanceHooks,
+    defaultRequestOptions: HttpRequestOptions,
+  ) {
+    this.defaultAxiosRequestConfig = defaultAxiosRequestConfig;
+    this.defaultRequestOptions = defaultRequestOptions;
+    this.axiosInstanceHooks = axiosInstanceHooks;
+    this.axiosInstance = this.createAxiosInstance(defaultAxiosRequestConfig);
     this.setupInterceptors();
   }
 
-  private createAxiosInstance(config: RawAxiosRequestConfig): AxiosInstance {
+  private createAxiosInstance(config: AxiosRequestConfig): AxiosInstance {
     return axios.create(config);
   }
 
-  private getAxiosConfig(): RawAxiosRequestConfig {
-    return this.axiosConfig;
+  private getDefaultAxiosRequestConfig(): AxiosRequestConfig {
+    return this.defaultAxiosRequestConfig;
+  }
+  private getDefaultRequestOptions(): HttpRequestOptions {
+    return this.defaultRequestOptions;
   }
 
-  private getAxiosHook(): AxiosHook {
-    return this.axiosHook;
+  private getAxiosInstanceHooks(): AxiosInstanceHooks {
+    return this.axiosInstanceHooks;
   }
 
   private getAxiosInstance(): AxiosInstance {
     return this.axiosInstance;
   }
 
-  private getDefaultRequestOptions(): RequestOptions {
-    return this.defaultRequestOptions;
-  }
-
-  private getAxiosHeaderStrategy(contentType: ContentTypeEnum): AxiosHeaderStrategy {
+  private createHttpHeaderPolicy<D = any>(contentType: ContentTypeEnum): HttpHeaderPolicy {
     switch (contentType) {
       case ContentTypeEnum.URL_ENCODED:
         return {
           headers: {
             'Content-Type': 'application/x-www-form-urlencoded',
           },
-          dataConvert: (data: unknown) => {
+          dataConvert: (data: D) => {
             return qs.stringify(data, { arrayFormat: 'brackets' });
           },
         };
       case ContentTypeEnum.MULTI_PART:
         return {
           headers: { 'Content-Type': 'multipart/form-data' },
-          dataConvert: (data: unknown) => {
+          dataConvert: (data: D) => {
             return data;
           },
         };
       default:
         return {
           headers: { 'Content-Type': 'application/json' },
-          dataConvert: (data: unknown) => {
+          dataConvert: (data: D) => {
             return JSON.stringify(data);
           },
         };
@@ -87,9 +90,9 @@ export class Axios {
   }
 
   private setupInterceptors(): void {
-    const hook = this.getAxiosHook();
+    const instanceHooks = this.getAxiosInstanceHooks();
 
-    if (!hook) {
+    if (!instanceHooks) {
       return;
     }
 
@@ -98,7 +101,7 @@ export class Axios {
       requestInterceptorsError,
       responseInterceptors,
       responseInterceptorsError,
-    } = hook;
+    } = instanceHooks;
 
     const axiosCanceler = new AxiosCanceler();
 
@@ -132,21 +135,27 @@ export class Axios {
 
   /**
    * 把当前请求的 options 与全局 options 整合获得一个完整的 options
+   * @param currentRequestOptions 当前请求的 options
+   * @returns 合并后的 options
    */
-  private mergeRequestOptions(options?: RequestOptions): RequestOptions {
-    const requestOptions = this.getDefaultRequestOptions();
-    if (!lodash.isEmpty(options)) {
-      return Object.assign({}, requestOptions, options);
+  private mergeHttpRequestOptions(currentRequestOptions?: HttpRequestOptions): HttpRequestOptions {
+    const defaultRequestOptions = this.getDefaultRequestOptions();
+    if (!lodash.isEmpty(currentRequestOptions)) {
+      return lodash.merge({}, defaultRequestOptions, currentRequestOptions);
     } else {
-      return requestOptions;
+      return defaultRequestOptions;
     }
   }
 
   /**
    * 把当前请求的 AxiosRequestConfig 与全局 AxiosRequestConfig 整合获得一个完整的 AxiosRequestConfig
+   * @param currentAxiosRequestConfig 当前请求的 AxiosRequestConfig
+   * @returns 合并后的 AxiosRequestConfig
    */
-  private mergeRequestConfigs<D = any>(config?: RawAxiosRequestConfig<D>): RawAxiosRequestConfig {
-    const axiosConfig = this.getAxiosConfig();
+  private mergeAxiosRequestConfigs<D = unknown>(
+    currentAxiosRequestConfig?: AxiosRequestConfig<D>,
+  ): AxiosRequestConfig<D> {
+    const defaultAxiosRequestConfig = this.getDefaultAxiosRequestConfig();
 
     const paramsSerializer: ParamsSerializerOptions = {
       serialize(params: any): string {
@@ -156,63 +165,134 @@ export class Axios {
       },
     };
 
-    const requestConfigs = Object.assign({ paramsSerializer }, axiosConfig);
+    const result = lodash.merge({ paramsSerializer }, defaultAxiosRequestConfig);
 
-    if (config) {
-      return Object.assign({}, requestConfigs, config);
+    if (!lodash.isEmpty(currentAxiosRequestConfig)) {
+      return lodash.merge({}, result, currentAxiosRequestConfig);
     } else {
-      return requestConfigs;
+      return result;
     }
   }
 
   private setupRequestStrategy<D = any>(
     url: string,
-    options: RequestOptions,
-    config?: RawAxiosRequestConfig<D>,
-  ): AxiosRequestStrategy {
-    const { onRequestHook } = this.getAxiosHook();
+    currentAxiosRequestConfig: AxiosRequestConfig<D>,
+    currentHttpRequestOptions: HttpRequestOptions,
+  ): HttpRequestPolicy {
+    const { onRequestHook } = this.getAxiosInstanceHooks();
 
     // 合并 options。把当前请求的 options 与全局 options 整合获得一个完整的 options
-    const requestOptions = this.mergeRequestOptions(options);
+    const httpRequestOptions = this.mergeHttpRequestOptions(currentHttpRequestOptions);
 
     // 合并 axios request config。把当前请求的 AxiosRequestConfig 与全局 AxiosRequestConfig 整合获得一个完整的 AxiosRequestConfig
-    let axiosRequestConfig: RawAxiosRequestConfig = this.mergeRequestConfigs<D>(config);
+    let axiosRequestConfig: AxiosRequestConfig<D> =
+      this.mergeAxiosRequestConfigs(currentAxiosRequestConfig);
     if (onRequestHook && lodash.isFunction(onRequestHook)) {
-      // 允许在 beforeRequestHook 中，对 AxiosRequestConfig 进行一些额外的设置
-      axiosRequestConfig = onRequestHook(axiosRequestConfig, requestOptions);
+      // 允许在 onRequestHook 中，对 AxiosRequestConfig 进行一些额外的设置
+      axiosRequestConfig = onRequestHook(axiosRequestConfig, httpRequestOptions);
     }
 
-    const contentType: ContentTypeEnum = requestOptions.contentType;
-    const strategy = this.getAxiosHeaderStrategy(contentType);
+    const httpHeaderPolicy = this.createHttpHeaderPolicy(httpRequestOptions.contentType);
+
     if (axiosRequestConfig.headers) {
-      axiosRequestConfig.headers = Object.assign(axiosRequestConfig.headers, strategy.headers);
+      axiosRequestConfig.headers = lodash.merge(
+        axiosRequestConfig.headers,
+        httpHeaderPolicy.headers,
+      );
     } else {
-      axiosRequestConfig.headers = strategy.headers;
+      axiosRequestConfig.headers = httpHeaderPolicy.headers;
     }
 
     axiosRequestConfig.url = url;
     if (!lodash.isEmpty(axiosRequestConfig.data)) {
-      axiosRequestConfig.data = strategy.dataConvert(axiosRequestConfig.data);
+      axiosRequestConfig.data = httpHeaderPolicy.dataConvert(axiosRequestConfig.data);
     }
 
     return {
       config: axiosRequestConfig,
-      options: requestOptions,
-      dataConvert: strategy.dataConvert,
+      options: httpRequestOptions,
     };
   }
 
+  /**
+   * 请求核心方法
+   * @param config axios request 必要参数
+   * @param options 针对每个请求特别指定的参数
+   * @returns 响应数据
+   */
+  public request<T = any, D = any>(
+    config: AxiosRequestConfig<D>,
+    options?: HttpRequestOptions,
+  ): Promise<AxiosHttpResult<T>> {
+    return new Promise<AxiosHttpResult<T>>((resolve, reject) => {
+      const { onResponseErrorHook, onResponseSuccessHook } = this.getAxiosInstanceHooks();
+      this.getAxiosInstance()
+        .request<HttpResult<T>, AxiosResponse<HttpResult<T>>, D>(config)
+        .then((response: AxiosResponse<HttpResult<T>>) => {
+          if (onResponseSuccessHook && lodash.isFunction(onResponseSuccessHook)) {
+            const result = onResponseSuccessHook(response, options);
+            resolve(result);
+          } else {
+            resolve(response);
+          }
+        })
+        .catch((error: AxiosError) => {
+          if (onResponseErrorHook && lodash.isFunction(onResponseErrorHook)) {
+            const result = onResponseErrorHook(error, options);
+            reject(result);
+          } else {
+            reject(error);
+          }
+        });
+    });
+  }
+
+  /**
+   * 处理请求。提取公共代码，避免每个方法中都要写一遍
+   * @param url 请求地址
+   * @param config axios request 必要参数
+   * @param options 针对每个请求特别指定的参数
+   * @returns 响应数据
+   */
+  private process<T = any, D = any>(
+    url: string,
+    config: AxiosRequestConfig<D>,
+    options = {} as HttpRequestOptions,
+  ): Promise<AxiosHttpResult<T>> {
+    let strategy = this.setupRequestStrategy<D>(url, config, options);
+    return this.request<T, D>(strategy.config, strategy.options);
+  }
+
+  /**
+   * GET
+   *
+   * <T> 返回响应中实际 data 中的内容类型
+   * <D> RequestBody 中的数据类型，实际对应 axios config 中的 data
+   *
+   * @param url 请求地址
+   * @param params 拼接在请求地址路径后面的参数，根据实际情况也可能不需要。
+   * @param options 对当前请求设置的参数
+   * @returns
+   */
   public get<T = any, D = any>(
     url: string,
     params = {},
     options = { contentType: ContentTypeEnum.JSON },
   ): Promise<AxiosHttpResult<T>> {
-    let policy = this.setupRequestStrategy<D>(url, options, { params, method: HttpMethodEnum.GET });
-    return this.request<T, D>(policy.config, policy.options);
+    return this.process<T, D>(
+      url,
+      {
+        params,
+        method: HttpMethodEnum.GET,
+      },
+      options,
+    );
   }
 
   /**
    * POST
+   *
+   * 针对 url 中有参数同时 request body 中也有数据的情况。额外增加一个方法，以防对现有的代码产生影响。
    *
    * <T> 返回响应中实际 data 中的内容类型
    * <D> RequestBody 中的数据类型，实际对应 axios config 中的 data
@@ -220,21 +300,24 @@ export class Axios {
    * @param url 请求地址
    * @param data 放置在 RequestBody 中的数据
    * @param options 对当前请求设置的参数。
-   * @param config 当前请求对 axios 特殊设置
+   * @param config 当前请求对 axios 特殊设置。POST 操作会有不同的 Content Type，以及不同的设置。
    * @returns
    */
   public post<T = any, D = any>(
     url: string,
     data: D,
     options = { contentType: ContentTypeEnum.JSON },
-    config?: RawAxiosRequestConfig<D>,
+    config?: AxiosRequestConfig<D>,
   ): Promise<AxiosHttpResult<T>> {
-    let policy = this.setupRequestStrategy<D>(url, options, {
-      ...config,
-      data,
-      method: HttpMethodEnum.POST,
-    });
-    return this.request<T, D>(policy.config, policy.options);
+    return this.process<T, D>(
+      url,
+      {
+        ...config,
+        data,
+        method: HttpMethodEnum.POST,
+      },
+      options,
+    );
   }
 
   /**
@@ -251,22 +334,25 @@ export class Axios {
    */
   public postWithParams<T = any, D = any>(
     url: string,
-    params = {},
+    params: Record<string, any>,
     data = {} as D,
     options = { contentType: ContentTypeEnum.JSON },
-    config?: RawAxiosRequestConfig<D>,
+    config?: AxiosRequestConfig<D>,
   ): Promise<AxiosHttpResult<T>> {
-    let policy = this.setupRequestStrategy<D>(url, options, {
-      ...config,
-      params,
-      data,
-      method: HttpMethodEnum.POST,
-    });
-    return this.request<T, D>(policy.config, policy.options);
+    return this.process<T, D>(
+      url,
+      {
+        ...config,
+        params,
+        data,
+        method: HttpMethodEnum.POST,
+      },
+      options,
+    );
   }
 
   /**
-   * 更新操作。
+   * PUT。更新操作。
    *
    * 针对 url 中有参数同时 request body 中也有数据的情况。额外增加一个方法，以防对现有的代码产生影响。
    *
@@ -283,18 +369,21 @@ export class Axios {
     url: string,
     data: D,
     options = { contentType: ContentTypeEnum.JSON },
-    config?: RawAxiosRequestConfig<D>,
+    config?: AxiosRequestConfig<D>,
   ): Promise<AxiosHttpResult<T>> {
-    let policy = this.setupRequestStrategy<D>(url, options, {
-      ...config,
-      data,
-      method: HttpMethodEnum.PUT,
-    });
-    return this.request<T, D>(policy.config, policy.options);
+    return this.process<T, D>(
+      url,
+      {
+        ...config,
+        data,
+        method: HttpMethodEnum.PUT,
+      },
+      options,
+    );
   }
 
   /**
-   * 更新操作。
+   * PUT。更新操作。
    *
    * 针对 url 中有参数同时 request body 中也有数据的情况。额外增加一个方法，以防对现有的代码产生影响。
    *
@@ -309,22 +398,25 @@ export class Axios {
    */
   public putWithParams<T = any, D = any>(
     url: string,
-    params = {},
+    params: Record<string, any>,
     data = {} as D,
     options = { contentType: ContentTypeEnum.JSON },
-    config?: RawAxiosRequestConfig<D>,
+    config?: AxiosRequestConfig<D>,
   ): Promise<AxiosHttpResult<T>> {
-    let policy = this.setupRequestStrategy<D>(url, options, {
-      ...config,
-      params,
-      data,
-      method: HttpMethodEnum.PUT,
-    });
-    return this.request<T, D>(policy.config, policy.options);
+    return this.process<T, D>(
+      url,
+      {
+        ...config,
+        params,
+        data,
+        method: HttpMethodEnum.PUT,
+      },
+      options,
+    );
   }
 
   /**
-   * 删除操作
+   * DELETE。删除操作
    *
    * <T> 返回响应中实际 data 中的内容类型
    * <D> RequestBody 中的数据类型，实际对应 axios config 中的 data
@@ -340,15 +432,18 @@ export class Axios {
     data = {} as D,
     options = { contentType: ContentTypeEnum.JSON },
   ): Promise<AxiosHttpResult<T>> {
-    let policy = this.setupRequestStrategy<D>(url, options, {
-      data,
-      method: HttpMethodEnum.DELETE,
-    });
-    return this.request<T, D>(policy.config, policy.options);
+    return this.process<T, D>(
+      url,
+      {
+        data,
+        method: HttpMethodEnum.DELETE,
+      },
+      options,
+    );
   }
 
   /**
-   * 删除操作。
+   * DELETE。删除操作
    *
    * 针对 url 中有参数同时 request body 中也有数据的情况。额外增加一个方法，以防对现有的代码产生影响。
    *
@@ -363,47 +458,18 @@ export class Axios {
    */
   public deleteWithParams<T = any, D = any>(
     url: string,
-    params = {},
+    params: Record<string, any>,
     data = {} as D,
     options = { contentType: ContentTypeEnum.JSON },
   ): Promise<AxiosHttpResult<T>> {
-    let policy = this.setupRequestStrategy<D>(url, options, {
-      params,
-      data,
-      method: HttpMethodEnum.DELETE,
-    });
-    return this.request<T, D>(policy.config, policy.options);
-  }
-
-  /**
-   * 请求核心方法
-   * @param config axios request 必要参数
-   * @param options 针对每个请求特别指定的参数
-   * @returns 响应数据
-   */
-  public request<T = any, D = any>(
-    config: RawAxiosRequestConfig<D>,
-    options?: RequestOptions,
-  ): Promise<AxiosHttpResult<T>> {
-    return new Promise<AxiosHttpResult<T>>((resolve, reject) => {
-      const { onResponseErrorHook, onResponseSuccessHook } = this.getAxiosHook();
-      this.getAxiosInstance()
-        .request<HttpResult<T>, AxiosResponse<HttpResult<T>>, D>(config)
-        .then((response: AxiosResponse<HttpResult<T>>) => {
-          if (onResponseSuccessHook && lodash.isFunction(onResponseSuccessHook)) {
-            const result = onResponseSuccessHook(response, options);
-            resolve(result);
-          } else {
-            resolve(response);
-          }
-        })
-        .catch((error: AxiosError) => {
-          if (onResponseErrorHook && lodash.isFunction(onResponseErrorHook)) {
-            reject(onResponseErrorHook(error, options));
-          } else {
-            reject(error);
-          }
-        });
-    });
+    return this.process<T, D>(
+      url,
+      {
+        params,
+        data,
+        method: HttpMethodEnum.DELETE,
+      },
+      options,
+    );
   }
 }